\name{AdaptLGL}
\alias{AdaptLGL}

\title{AdaptLGL}
\description{    
adapt interactively a Gaussians Mixture Model GMM to the empirical PDF of the data such that N(M,S)*W is a  model for Data 
}
\usage{
AdaptLGL(Data,Means=NaN,SDs=NaN,Weights=NaN,

ParetoRadius=NaN,LB=NaN,HB=NaN,ListOfAdaptGauss)
}

\arguments{
  \item{Data}{ Data for empirical PDF. Has to be an Array of values. NaNs
  and NULLs will be deleted }
  \item{Means}{[1:3] Optional: Means of gaussians of GMM. Mean[2] will be a Gauss, Mean[1] and Mean[3] a logGauss. }
  \item{SDs}{[1:3] Optional: Deviations of gaussians of GMM. (Has to be the same length as Means) }
  \item{Weights}{[1:3] Optional: Weights of gaussians of GMM. (Has to be the same length as Means) }
  \item{ParetoRadius}{ Optional: Pareto Radius of Pareto Desity Estimation (PDE). }
  \item{LB}{ Optional: Low boundary of estimation. All values below LB will be deleted. Default: min(Data) }
  \item{HB}{ Optional: High boundary of estimation. All values above HB will be deleted. Default: max(Data) }
    \item{ListOfAdaptGauss}{ ...}
}
\details{
Adapt interactively a Gaussians Mixture Model GMM (logGauss - Gauss - logGauss) to the empirical PDF
  of the data (generated by ParetoDensityEstimation) such that N(M,S)*W is a  model for Data
	
    Data: maximum length is 10000. If larger, Data will be randomly reduced to 10000 Elements.
  MeansIn/DeviationsIn/WeightsIN: If empty, 3 gaussians (2 logGauss and 1 Gauss, in order logGauss - Gauss - logGauss) are generated.
  Pareto Radius: If empty: will be generated by ParetoDensityEstimation
}
\value{
List with
  \item{Means}{Means of gaussians.}
  \item{SDs}{Standard Deviations of gaussians.}
  \item{Weights}{Weights of gaussians.}
  \item{ParetoRadius}{Pareto Radius: Either ParetoRadiusIn, the pareto radius enerated by PretoDensityEstimation(if no Pareto Radius in Input).}
  \item{RMS}{Root Mean Sqare of Daviation between Gaussians Mixture Model GMM to the empirical PDF. Normalized by RMS of one gaussian with mean=meanrobust(data) and sdev=stdrobust(data).}
}

\author{ Onno Hansen-Goos }

\examples{
  data1=c(rnorm(1000))
  \dontrun{Vals1=AdaptGauss(data1)}
  
  data2=c(rnorm(1000),rlnorm(2000)+2,-rlnorm(1000)*2-1)
  \dontrun{Vals2=AdaptLGL(data2,c(-2,0,4),c(2,1,1),c(0.25,0.25,0.5),0.3,-6,6)}

 }

