\name{EM_LGL}
\alias{EM_LGL}
\title{EM_LGL}
\description{
EM_LGL
}
\usage{
EM_LGL(Data, Means, Sdevs, alphas, MaxNumberofIterations)
}
\arguments{
  \item{Data}{  the data points as column vactor}
  \item{Means}{  estimated Kernels = means(clustercenters), }
  \item{Sdevs}{ estimated Kernels = standard deviations}
  \item{alphas}{                relative number of points in Pdfs (prior probabilities): sum(alphas) ==1
  the second number of the Vectors belongs to the Gaus distribution,
    the third number of the Vectors belongs to the positiv Log Gaus distribution.
  
  }
  \item{MaxNumberofIterations}{ iterations of em }
}

\value{

   OUTPUT
   EMmean(1:3)                     means of LGL generated by EM algorithm
   EMsdev(1:3)                     standard deviations of LGL generated by EM algorithm
   EMalphas(1:3)                    prior probabilities of Pdfs
}
\author{Florian Meyer}

\keyword{LGL}
\keyword{ LogGaussLog}
