% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Intersect2Mixtures.R
\name{Intersect2Mixtures}
\alias{Intersect2Mixtures}
\title{Intersect of two Gaussians}
\usage{
Intersect2Mixtures(Mean1, SD1, Weight1, Mean2, SD2, Weight2,
  IsLogDistribution = c(FALSE, FALSE), MinData, MaxData)
}
\arguments{
\item{Mean1}{mean of 1.gaussian}

\item{SD1}{standard deviations of 1.gaussian}

\item{Weight1}{weight of 1. guassian}

\item{Mean2}{mean of 2.gaussian}

\item{SD2}{standard deviations of 2.gaussian}

\item{Weight2}{weight of 2. guassian}

\item{IsLogDistribution}{Optional, ==1 if distribution(i) is a LogNormal, default vector of zeros of length 2}

\item{MinData}{Optional, Beginning of range, where the intersect is searched for, default min(Mean1,Mean2)}

\item{MaxData}{Optional, End of range, where the intersect is searched for, default max(Mean1,Mean2)}
}
\value{
CutX x value, where gaussian 1=gaussian2

CutY y value, where gaussian 1=gaussian2
}
\description{
Finds the intersect of two gaussians or log gaussians
}
\author{
Michael Thrun, Rabea Griese

\strong{See Also}

BayesDecisionBoundaries
}
