\name{LKWFahrzeitSeehafen2010}
\alias{LKWFahrzeitSeehafen2010}
\docType{data}
\title{
Truck driving time seaport 2010
}
\description{
Truck driving time to seaports measured in 2010.
}
\usage{data("LKWFahrzeitSeehafen2010")}
\format{
  The format is:
 num [1:11441] 84.7 13.2 11.5 41.4 52.9 ...
}

\references{
Behnisch, M., Ultsch, A.: Knowledge Discovery in Spatial Planning Data - A Concept for Cluster Understanding, in: Helbich, M., Arsanjani, J. J., Leitner, M. (eds.): Computational Approaches for Urban Environments, in: Gatrell, J.D., Jensen, R.R.: Geotechnologies and the Environment Series, Vol, 13, Springer, Berlin, pp. 49-75, 2015.
}
\examples{
data(LKWFahrzeitSeehafen2010)
## maybe str(LKWFahrzeitSeehafen2010) ; plot(LKWFahrzeitSeehafen2010) ...
}
\keyword{datasets}
