\name{AdaptGauss-package}
\alias{AdaptGauss-package}
\alias{MultiModal-package}
\alias{MultiModal}
\docType{package}
\title{
AdaptGauss-package
}
\description{
Multimodal distributions can be modelled as a mixture of components. The model is derived using the Pareto Density Estimation (PDE) for an estimation of the pdf [Ultsch 2015]. PDE has been designed in particular to identify groups/classes in a dataset. The expectation maximization algorithm estimates a Gaussian mixture model of density states [Bishop 2006] and the limits between the different states are defined by Bayes decision boundaries [Duda 2001]. The model can be verified with Chi-squared test and QQ plot.
}

\author{
 Michael Thrun, Onno Hansen-Goos, Rabea Griese, Catharina Lippmann,Jorn Lotsch, Alfred Ultsch
 
 Maintainer: Michael Thrun <mthrun@mathematik.uni-marburg.de>
}
\references{
Alfred Ultsch, Michael C. Thrun, Onno Hansen-Goos, Joern Loetsch Identification of molecular fingerprints in heat pain thresholds by use of an interactive mixture model toolbox (AdaptGauss), IJMS 2015.

Duda, R.O., P.E. Hart, and D.G. Stork, Pattern classification. 2nd. Edition. New York, 2001, p 512 ff

Bishop, Christopher M. Pattern recognition and machine learning. springer, 2006, p 435 ff

Ultsch, A., Pareto Density Estimation: A Density Estimation for Knowledge Discovery, in Innovations in classification, data science, and information systems, Springer, New York, pp 91-100, 2005.

Thrun M.C.,Ultsch, A., Models of Income Distributions for Knowledge Discovery, European Conference on Data Analysis, DOI 10.13140/RG.2.1.4463.0244, Colchester 2015.
}

\keyword{Multimodal}
\keyword{AdaptGauss}
\keyword{EM}
\keyword{expectation maximization}
\keyword{pareto density estimation}
\keyword{pdf}
\keyword{mixture of components}
\keyword{GMM}
\keyword{gaussian mixture model}
\keyword{mixture}
