% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassification.R
\name{BayesClassification}
\alias{BayesClassification}
\title{BayesClassification}
\usage{
BayesClassification(Data, Means, SDs, Weights, IsLogDistribution = Means
  * 0, ClassLabels = c(1:length(Means)))
}
\arguments{
\item{Data}{vector of Data}

\item{Means}{vector[1:L] of Means of Gaussians (of GMM)}

\item{SDs}{vector of standard deviations, estimated Gaussian Kernels, has to be the same length as Means}

\item{Weights}{vector of relative number of points in Gaussians (prior probabilities), has to be the same length as Means}

\item{IsLogDistribution}{Optional, ==1 if distribution(i) is a LogNormal, default vector of zeros of length 1:L}

\item{ClassLabels}{Optional numbered class labels that are assigned to the classes. default (1:L), L number of different components of gaussian mixture model}
}
\value{
Cls(1:n,1:d) classiffication of Data, such that 1= first component of gaussian mixture model, 2= second component of gaussian mixture model and so on. For Every datapoint a number is returned.
}
\description{
Bayes Klassifikation den Daten zuordnen
}
\author{
Michael Thrun
}
